+++ main.py
@@ -2,8 +2,8 @@
 import argparse
 
 # from config import settings
-from src.link_records import LinkRecords
-from src.dedupe_records import DedupeRecords
+from src.models.link_records import LinkRecords
+from src.models.dedupe_records import DedupeRecords
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
