+++ src/handlers/streaming_xml_handler.py
