+++ src/ingest/marc_to_csv.py
@@ -3,7 +3,7 @@ import os.path
 from xml.sax import SAXParseException
 from pymarc import parse_xml_to_array
 from pymarc import parse_json_to_array
-from src.marc_record import MarcRecord
+from src.normalize.marc_record import MarcRecord
 
 
 class MarcToCsv:
