+++ tests/normalize/test_link_records_file.py
@@ -1,5 +1,5 @@
 import os
-from src.link_records_file import LinkRecordsFile
+from src.normalize.link_records_file import LinkRecordsFile
 
 
 def test_importing_xml_file(example_dictionary):
