+++ tests/scoring/test_score_model.py
@@ -1,4 +1,4 @@
-from src.score_model import ScoreModel
+from src.scoring.score_model import ScoreModel
 
 
 def test_scoring():
